Fix the CC-BY-CA logo format in html
CC-BY-SA logo is set to float right.
This works fine in PDF in most versions of the tools.
This works correctly in Ubuntu 20.04, Debian 9, and Fedora 34.
The float right even with out this patch is broken in Ubuntu 18.04 and
Debian 10 for PDF.

However in all versions the CC-BY-SA logo in HTML was not correct.  The
revision table started before the logo finished and thus was diminished
width for it's whole length.

This new ReST construct adds a <div class=clearer>&nbsp</div> before
the table.  div.clearer is style "clear:both" in _static/basic.css

Signed-off-by: Bill Mills <bill.mills@linaro.org>