Fix the CC-BY-CA logo format in html
CC-BY-SA logo is set to float right.
For PDF this works fine in most versions of the tools.
Known working: Ubuntu 20.04, Debian 9, and Fedora 34.
The float right even with out this patch is broken in Ubuntu 18.04 and
Debian 10 for PDF.

However in all versions, the CC-BY-SA logo in HTML was not correct.  The
revision table started before the logo finished and thus was diminished
width for it's whole length.

This injects raw html but only for html modes.
The html injected is an empty div with style="clear:both" before
the table.

Signed-off-by: Bill Mills <bill.mills@linaro.org>