add scripts/setup.sh for project automation
Add the scripts/setup.sh file to prepare the container/VM and
build the project.

Instead of embedding this in CI specific job file, put it here for everyone
to use.  The brave may even choose to run it on their non-VM machine.

Not all the allowed for distro's work.  See distro-build-status.

Signed-off-by: Bill Mills <bill.mills@linaro.org>